:recycle:대댓글 부분 수정

+++ src/main/java/com/prgrms/team03linkbookbe/like/repository/LikeRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
-    Long countByFolderEquals(Folder folder);
+    int countByFolderEquals(Folder folder);
 
     Optional<Like> findByIdAndUser(Long id, User user);
